plat: marvell: otx2: change disable_ooo setting as a core mask
disable_ooo setting in u-boot will not be considered a core mask
to selectively disable ooo on particular cores

Integrated from branch atf-v2.2-devel.
ORIGINAL: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/32663

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/boot/atf/+/33241
Tested-by: Chandrakala Chavva <cchavva@marvell.com>

+++ include/plat/marvell/octeontx/otx2/plat_otx2_configuration.h
@@ -31,7 +31,7 @@ int plat_octeontx_get_nix_count(void);
 int plat_octeontx_get_mcc_count(void);
 int plat_octeontx_get_pem_count(void);
 int plat_get_altpkg(void);
-int plat_get_ooo_status(void);
+uint64_t plat_get_ooo_status(void);
 unsigned int plat_configure_rid(void);
 unsigned int plat_configure_cpt_rid(void);
 struct otx2_stream_security_setting *plat_get_otx2_stream_security(int *count);
