+++ include/plat/marvell/octeontx/otx2/plat_svc.h
@@ -92,7 +92,7 @@ uintptr_t otx2_svc_smc_handler(uint32_t smc_fid,
 			       void *handle,
 			       u_register_t flags);
 
-int octeontx2_configure_ooo(int x1);
+int octeontx2_configure_ooo(uint64_t x1);
 
 #endif /* __PLAT_SVC_H__ */
